Update to 1.2.0
