+++ .gitignore
@@ -1,3 +1,2 @@
 /.idea
 /.vscode
-pvm.exe
